+++ nnringwidth/README.md
@@ -1,0 +1,4 @@
+# align_edgering_length
+<img width="598" alt="t" src="https://user-images.githubusercontent.com/159296/132953492-813bed4d-da0d-4fed-ac55-17033bf51811.png">
+
+A tool for Maya that unifies the width of polygons arranged in a strip and sets the taper numerically.
